pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pylibre"
version = "0.1.0"
description = "A Python client for interacting with the Libre blockchain"
readme = "README.md"
authors = [
    {name = "Benobi", email = "python@benobi.one"}
]
license = {text = "MIT"}
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]
dependencies = [
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "pyntelope>=0.2.0",
    "websockets>=12.0",
    "python-binance>=1.0.19",
]

[project.urls]
Homepage = "https://github.com/bensig/pylibre"
Repository = "https://github.com/bensig/pylibre.git"

[project.scripts]
pylibre = "pylibre.cli:main"

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi-line-output = 3

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--strict-markers --tb=short"
testpaths = [
    "tests",
]
pythonpath = [
    ".",
    "src"
]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
files = "src/pylibre"
warn_unused_configs = true
ignore_missing_imports = true
strict_optional = true
show_error_codes = true

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.5",
    "black>=24.1.1",
    "isort>=5.13.2",
    "mypy>=1.8.0",
    "types-requests>=2.31.0.20240125",
    "pyyaml>=6.0.2",
]
